Written function game and winCondition
In game function's body playRound function is called 5 times until
score of one of the variables reaches 5. In winCondition function
condition for the win is set when one of the variables equals 5.
Previously made function getComputerChoice and user input are nested
in the game function so the function can use them multiple times.
winCondition function is called once so it is nested in the game
function but outside while loop. Finally game function is called.